% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ss_catch.R
\name{compile_ss_catch}
\alias{compile_ss_catch}
\title{Compile catch data in Stock Synthesis format}
\usage{
compile_ss_catch(
  years,
  seas,
  fleet_number,
  catch,
  catch_se,
  initial_catch = 1e-20,
  initial_catch_se = 0.01
)
}
\arguments{
\item{seas}{An integer specifying the season associated with the
catch observations in Stock Synthesis.}

\item{fleet_number}{An integer specifying the Stock Synthesis fleet
number.}

\item{catch}{A numeric vector containing catch observations
corresponding to \code{year}.}

\item{catch_se}{A numeric value specifying the standard error
assigned to catch observations for all years in the time series.}

\item{initial_catch}{A numeric value specifying the initial
equilibrium catch for the fleet in Stock Synthesis.}

\item{initial_catch_se}{A numeric value specifying the standard
error of the initial equilibrium catch for the fleet.}

\item{year}{A numeric vector specifying the years in the catch
time series.}
}
\value{
A data frame containing the catch time series formatted
  for Stock Synthesis, including year, season, fleet number, catch,
  and catch standard error. The initial equilibrium catch and its
  standard error are included as specified.
}
\description{
Compiles a catch time series for a single fleet into the format
required by Stock Synthesis. The function assigns the year, season,
and fleet number to each catch observation and specifies the catch
standard error. An initial equilibrium catch and its standard error
can also be included.
}
\examples{
compile_ss_catch(
  year = c(1990, 1991, 1992, 1993, 1994, 1995),
  seas = 3,
  fleet_number = 2,
  catch = c(2234, 2342, 2253, 2332, 2354, 2300),
  catch_se = 0.01,
  initial_catch = 100,
  initial_catch_se = 0.5
)

}
