% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSHelper.R
\name{draw_image}
\alias{draw_image}
\title{Visualize}
\usage{
draw_image(cover_images, ...)
}
\arguments{
\item{x}{An \code{\link{SSHelper}} SSHelper object}
}
\description{
Given an \code{\link{SSHelper}} object, this \code{\link[base:draw_image]{base::draw_image}} method retrieve an image file associated with
the category type from Smith College library and displays it in the graphics device.
}
