% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSHelper.R
\name{library_search}
\alias{library_search}
\title{Retrieve metadata from Smith College library}
\usage{
library_search(search_terms, place_to_search, ...)
}
\arguments{
\item{search_terms}{user input specifying the library category of choice
books, articles, rare_books, video, journals, and special_collections}

\item{place_to_search}{user input specifying the title of the requested source}

\item{...}{Currently ignored}
}
\value{
A data frame of lists of requested catalog \code{library_search} with following fields
\itemize{
\item journals
\item books
\item articles
\item video
\item special_collections
\item rare_books
}
}
\description{
SSHelper is a package that assists searches in the Smith college library catalog
it allows users to browse lists of sources and cover images
}
