% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_cpue_multi.R
\name{loess_cpue_multi}
\alias{loess_cpue_multi}
\title{Fitting LOESS smoothing to abundance index}
\usage{
loess_cpue_multi(
  data,
  year_col = "Year",
  cpue_col = "CPUE",
  span_range = seq(0.4, 0.7, by = 0.1),
  chosen_span = 0.6,
  palette = "Dark 3",
  plot_results = TRUE
)
}
\arguments{
\item{data}{A data frame with year in first column and a relative abundance index (e.g. Catch per unit effort) in second column.}

\item{year_col}{Column name as a string indicating the year in 'data'.}

\item{cpue_col}{Column name as a string indicating the index in 'data'.}

\item{span_range}{A vector indicating a range of smoothing parameters to be applied  e.g., c (0.1, 0.4, 0.6)}

\item{chosen_span}{A value indicating the choice of smoothing parameter to append SEs to the data.}

\item{palette}{A colour theme as string for graphical display of smoothing parameters applied to the index.}

\item{plot_results}{TRUE or FALSE to indicate the choice of a graphical display.}
}
\value{
The temperature in degrees Celsius
}
\description{
Abundance indices in Stock Synthesis are typically standardised to remove effects that are not related to fishing and are assumed to have a log-normal error structure with units of se of log (index). Fitting a LOESS smoothing function is one way of explaining this variance.
}
\examples{
year<-c(2010:2024)

Std_CPUE<-c(0.25, 0.26, 0.24, 0.31, 0.28, 0.25, 0.23, 0.19,
            0.22, 0.35, 0.45, 0.41, 0.42, 0.49, 0.64)
            
Abundance <-data.frame(Year=year,Index=Std_CPUE)

loess_result <- loess_cpue_multi(
  data = Abundance,
  year_col = "Year",
  cpue_col = "Index",
  span_range = seq(0.6, 0.8, by = 0.1),
  chosen_span = 0.6
)

print(loess_result$updated_data)
}
