% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSHelper_data}
\alias{SSHelper_data}
\title{Minidata to show example of library search result}
\format{
A data frame with 6 rows and 2 columns
\describe{
\item{title}{books title (minumum 1, maximum 6)}
\item{image}{image URL}
}
@source \url{https://www.smith.edu/libraries/discover/search.php}
}
\usage{
SSHelper_data
}
\description{
Minidata to show example of library search result
}
\keyword{datasets}
