% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSHelper.R
\name{draw_image}
\alias{draw_image}
\title{Visualize}
\usage{
draw_image(search_results, ...)
}
\arguments{
\item{search_results}{An \code{\link{library_search}} SSHelper object}

\item{...}{Currently ignored}
}
\description{
Given an \code{\link{library_search}} object, this \code{\link{draw_image}} function retrieve an image file associated with
the category type from Smith College library and displays it in the graphics device.
}
