% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_timeseries.R
\name{fill_timeseries}
\alias{fill_timeseries}
\title{Function to fill discontinuous time series of catch with a user chosen value}
\usage{
fill_timeseries(catch.ts, fleetname, start.year, end.year, value)
}
\arguments{
\item{catch.ts}{A data frame with year in first column and catch in second column.}

\item{fleetname}{An integer or string indicating the fleet type in Stock Synthesis.}

\item{start.year}{An integer indicating the first year of the resultant time series.}

\item{end.year}{An integer indicating the last year of the resultant time series.}

\item{value}{User chosen value to fill the years where there is a missing catch.}
}
\value{
A dataframe with complete time series of year, fleet and catch
}
\description{
Years with missing catch will be filled with a value
}
\examples{
year<-c(1,3,5,8,14)
catch<-c(1000,500,300,400,140)
annual.catch<-data.frame(year,catch)

fill_timeseries(
                catch.ts  =annual.catch,
                fleetname =1,
                start.year=1,
                end.year  =20,
                value     =10
               )
}
