% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSHelper.R
\name{library_search}
\alias{library_search}
\title{Retrieve metadata from Smith College library}
\usage{
library_search(search_terms, place_to_search, ...)
}
\value{
A list of category \code{SSHelper} with following fields
\itemize{
\item journals
\item books
\item articles
\item video
\item special_collections
}
}
\description{
A short description...
}
