% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_cpue.R
\name{plot_catch_cpue}
\alias{plot_catch_cpue}
\title{Plot catch and CPUE}
\usage{
plot_catch_cpue(
  catch_data,
  cpue_data,
  catch_col = "catch",
  year_col = "year",
  cpue_year_col = "Year",
  cpue_index_col = "Index",
  x_tick = 2,
  catch_color = "blue",
  cpue_color = "red"
)
}
\arguments{
\item{catch_data}{A data frame containing catch data.}

\item{cpue_data}{A data frame containing CPUE index data.}

\item{catch_col}{Character string giving the name of the catch column.}

\item{year_col}{Character string giving the name of the year column.}

\item{cpue_year_col}{Character string giving the name of the year
column in `cpue_data`.}

\item{cpue_index_col}{Character string giving the name of the CPUE
index column.}

\item{x_tick}{Numeric value specifying the interval between x-axis
breaks.}

\item{catch_color}{Character string specifying the colour for catch.}

\item{cpue_color}{Character string specifying the colour for CPUE.}
}
\value{
A ggplot object showing total catch and CPUE.
}
\description{
Combines catch across fleets and plots total catch against a CPUE
index using a secondary y-axis.
}
\examples{
# Example: Catch and CPUE time series

# Create catch data for multiple fleets from 1950 to 2022
set.seed(124)

years <- 1950:2022

catch_data <- data.frame(
  year = rep(years, each = 2),
  fleet = rep(c("Fleet 1", "Fleet 2"), length(years))
)

# Simulate catch with increasing catch through time
catch_data$catch <- ifelse(
  catch_data$year < 1980,
  runif(nrow(catch_data), 5, 30),
  ifelse(
    catch_data$year < 1995,
    runif(nrow(catch_data), 30, 150),
    runif(nrow(catch_data), 150, 500)
  )
)


# Create CPUE data for the later part of the time series
cpue_years <- 2005:2022

cpue_data <- data.frame(
  Year = cpue_years,
  Index = c(
    0.16, 0.17, 0.14, 0.15, 0.13, 0.10,
    0.11, 0.15, 0.14, 0.10, 0.11, 0.12,
    0.16, 0.20, 0.13, 0.15, 0.12, 0.14
  )
)

# Plot catch and CPUE
plot_catch_cpue(
  catch_data = catch_data,
  cpue_data = cpue_data,
  x_tick = 4
)
}
